requirements.txt
auditlog3==1.0.1
dj-database-url==0.5.0
dj-static==0.0.6
django-extensions==3.1.3
django-localflavor==3.1
django-ninja==0.13.2
django-utils-six==2.0
django-widget-tweaks==1.4.8
Django==3.2.*
Faker==8.8.1
pytest-django==4.4.0
python-decouple==3.4
